+++ libs/oidc/src/index.ts
@@ -6,3 +6,4 @@ export * from './guards';
 export * from './interfaces';
 export * from './strategies';
 export * from './utils';
+export * from './middlewares';